+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/markets/BaseBet.java
@@ -129,5 +129,5 @@ public abstract class BaseBet {
     /**
      * DI prediction metadata for the bet.
      */
-    @Nullable public PredictionData predictionData;
+    @Nullable public BetPredictionData predictionData;
 }