+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/markets/Market.java
@@ -46,5 +46,5 @@ public class Market {
     /**
      * DI prediction metadata for the market.
      */
-    @Nullable public PredictionData predictionData;
+    @Nullable public MarketPredictionData predictionData;
 }