+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/markets/ProviderMarket.java
@@ -41,5 +41,5 @@ public class ProviderMarket {
     /**
      * DI prediction metadata for the provider market.
      */
-    @Nullable public PredictionData predictionData;
+    @Nullable public MarketPredictionData predictionData;
 }